feat:강의내용 따라하기(Resource, Profile)
