+++ src/main/java/org/prgrms/kdtspringdemo/voucher/JdbcVoucherRepository.java
@@ -1,6 +1,7 @@
 package org.prgrms.kdtspringdemo.voucher;
 
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
 
 import java.util.*;
@@ -8,7 +9,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Stream;
 
 @Repository
-@Qualifier("jdbc")
+@Profile("dev")
 public class JdbcVoucherRepository implements VoucherRepository {
     private final Map<UUID, Voucher> storage = new ConcurrentHashMap<>();
