+++ src/main/resources/application.yaml
@@ -1,3 +1,8 @@
+server:
+  - dev.bar.com
+  - foo.bar.com
+---
+spring.config.activate.on-profile: dev
 kdt:
   version: "v1.0"
   minimum-order-amount: 1
@@ -10,4 +15,17 @@ kdt:
     line 1 hello world
     line 2 xxxx
     line 3
-
+---
+spring.config.activate.on-profile: local
+kdt:
+  version: "v1.0"
+  minimum-order-amount: 1
+  support-vendors:
+    - a
+    - b
+    - c
+    - d
+  description: |-
+    line 1 hello world
+    line 2 xxxx
+    line 3
