+++ test/sample.txt
@@ -1,0 +1,2 @@
+hello I'm a sample file!
+Hi hi hmmm...
